+++ lib/recurly/client.php
@@ -29,7 +29,7 @@ class Recurly_Client
    */
   private $_acceptLanguage = 'en-US';
 
-  const API_CLIENT_VERSION = '2.0.7';
+  const API_CLIENT_VERSION = '2.1.0';
   const DEFAULT_ENCODING = 'UTF-8';
 
   const GET = 'GET';
@@ -49,6 +49,8 @@ class Recurly_Client
   const PATH_PLANS = '/plans';
   const PATH_SUBSCRIPTIONS = '/subscriptions';
   const PATH_TRANSACTIONS = '/transactions';
+
+  const PATH_RECURLY_JS_RESULT = '/recurly_js/result';
 
   const PATH_TRANSPARENT = '/transparent/';
   const PATH_TRANSPARENT_SUBSCRIPTION = '/subscription';